refactor: rename macro.internal.ts to _macro.ts
• Move macro.internal.ts to _macro.ts for better file organization

• Update imports in _macro.ts and pricing-fetcher.ts

• Follow naming convention of underscore prefix for internal files
